remove required_status_checks for make, for now
These required_status_checks apply for every branch, which breaks
backports when the job cannot run and succeed. Since the make flow
wasn't available on older branches it cannot pass on those branches.

We remove them for now until the required_status_checks are made
specific to branches where they can succeed. We'll have the opposite
problem to this down the road - we'll want to remove the bazel checks,
but release-1.9 probably won't have bazel.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>