test: using cron periodics rather than intervals
If we move to generating our tests, we can have our generator
automatically spread periodics throughout the hour using cron-style
definitions intead of intervals.

This should ensure that we avoid situations where we have a tonne of
tests which all run on the hour causing a spike in resources but
leaving the cluster less utilized the rest of the time.

This is a manual test to check that cron-style jobs work on our deployed
version of prow today.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>